+++ tests/test_io.py
@@ -1,7 +1,7 @@
 """
 Test file input/output.
 """
-
+import os
 import unittest
 import numpy as np
 
@@ -9,6 +9,9 @@ import ams
 
 
 class TestMATPOWER(unittest.TestCase):
+    """
+    Test IO functions for MATPOWER and PYPOWER.
+    """
 
     def setUp(self):
         self.mpc5 = ams.io.matpower.m2mpc(ams.get_case('matpower/case5.m'))
@@ -34,6 +37,12 @@ class TestMATPOWER(unittest.TestCase):
         np.testing.assert_array_equal(system5.GCost.c2.v,
                                       np.zeros(system5.StaticGen.n))
 
+        # Check if Area is added
+        self.assertGreater(system5.Area.n, 0)
+
+        # Check if Zone is added
+        self.assertGreater(system5.Zone.n, 0)
+
         system14 = ams.system.System()
         # Test gentype length check
         mpc14 = self.mpc14.copy()
@@ -46,6 +55,35 @@ class TestMATPOWER(unittest.TestCase):
         np.testing.assert_array_less(np.zeros(system14.StaticGen.n),
                                      system14.GCost.c2.v,)
 
+    def test_system2mpc(self):
+        """Test conversion from AMS System to MPC."""
+        system5 = ams.system.System()
+        ams.io.matpower.mpc2system(self.mpc5, system5)
+        mpc5 = ams.io.matpower.system2mpc(system5)
+
+        self.assertEqual(mpc5['baseMVA'], self.mpc5['baseMVA'])
+
+        # Bus
+        # type, PD, QD, GS,BS, VM, VA. BASE_KV, VMAX, VMIN
+        bus_cols = [1, 2, 3, 4, 5, 7, 8, 9, 11, 12]
+        np.testing.assert_array_equal(mpc5['bus'][:, bus_cols],
+                                      self.mpc5['bus'][:, bus_cols])
+
+        # Branch, Gen, Gencost, can have minor differences but is okay
+
+        # String type data
+        np.testing.assert_array_equal(mpc5['gentype'], self.mpc5['gentype'])
+        np.testing.assert_array_equal(mpc5['genfuel'], self.mpc5['genfuel'])
+        np.testing.assert_array_equal(mpc5['bus_name'], self.mpc5['bus_name'])
+
+        # Area quantity
+        self.assertEqual(np.unique(mpc5['bus'][:, 6]).shape[0],
+                         np.unique(self.mpc5['bus'][:, 6]).shape[0])
+
+        # Zone quantity
+        self.assertEqual(np.unique(mpc5['bus'][:, 10]).shape[0],
+                         np.unique(self.mpc5['bus'][:, 10]).shape[0])
+
     def test_gencost1(self):
         """Test when gencost is type 1."""
         mpcgc1 = self.mpc14.copy()
@@ -54,3 +92,39 @@ class TestMATPOWER(unittest.TestCase):
         system = ams.system.System()
         ams.io.matpower.mpc2system(mpcgc1, system)
         self.assertEqual(system.GCost.n, 5)
+
+    def test_mpc2m(self):
+        """Test conversion from MPC to M file."""
+        mpc5 = ams.io.matpower.m2mpc(ams.get_case('matpower/case5.m'))
+        mpc14 = ams.io.matpower.m2mpc(ams.get_case('matpower/case14.m'))
+
+        # Test conversion to M file
+        mfile5 = ams.io.matpower.mpc2m(mpc5, './case5out.m')
+        mfile14 = ams.io.matpower.mpc2m(mpc14, './case14out.m')
+
+        # Check if the files exist
+        self.assertTrue(os.path.exists(mfile5))
+        self.assertTrue(os.path.exists(mfile14))
+
+        mpc5read = ams.io.matpower.m2mpc(mfile5)
+        mpc14read = ams.io.matpower.m2mpc(mfile14)
+
+        # Check if the numerical values are the same
+        for key in mpc5:
+            if key in ['bus_name', 'gentype', 'genfuel']:
+                continue
+            np.testing.assert_array_almost_equal(
+                mpc5[key], mpc5read[key], decimal=5,
+                err_msg=f"Mismatch in {key} when converting case5.m"
+            )
+        for key in mpc14:
+            if key in ['bus_name', 'gentype', 'genfuel']:
+                continue
+            np.testing.assert_array_almost_equal(
+                mpc14[key], mpc14read[key], decimal=5,
+                err_msg=f"Mismatch in {key} when converting case14.m"
+            )
+
+        # Clean up the generated files
+        os.remove(mfile5)
+        os.remove(mfile14)